+++ themes/default/views/layouts/fullmenu.php
@@ -296,6 +296,14 @@ $cs->registerCssFile(Yii::app()->baseUrl . "/sass/css/main.css?v=" . TAG_VERSION
                                         </a>
                                     </li>
                                 <?php endif ?>
+                                <?php if (Yii::app()->features->isEnable("FEAT_GRADESRELEASE")): ?>
+                                    <li class="t-menu-item <?= strpos($_SERVER['REQUEST_URI'], "?r=enrollment/gradesRelease") ? 'active' : '' ?>">
+                                        <a class="t-menu-item__link" href="<?php echo yii::app()->createUrl('enrollment/gradesRelease') ?> ">
+                                            <span class="t-report_card t-menu-item__icon"></span>
+                                            <span class="t-menu-item__text">Lançamento de Notas</span>
+                                        </a>
+                                    </li>
+                                <?php endif ?>
                             </ul>
                         </li>
                         <?php if (Yii::app()->getAuthManager()->checkAccess('instructor', Yii::app()->user->loginInfos->id)) : ?>